Feat: VP parser
This branch is an extension on vc-issuance previous work.
Currently, vc-issuance can be used in the service-offering wizzard because of previous format change, which analyse the first VC of the VP input.
This extension, vc-issuance-experimental, uses a graph to parse the VC payload, removing the hard coded path of the self description VC.
It also allows the compliance service to verify shape and content of RegistrationNumber VC, T&C VC and Participant Credential VC (which needs to be implemented)
Further testing are needed on this new route, especially on the service-offering part, which need an online deployment.
Previous vc-issuance route is to be deprecated once the wizzard will be able to connect to the new one.
This route is still WIP and may need some adjustment later.

Signed-off-by: Henry Faure-Geors <henry.faure-geors@atos.net>