Make the did.json W3C compliant
Description: Currently the did.json located in /.well-known/did.json is not W3C compliant. It is missing 'type' and 'controller' according to https://www.w3.org/TR/did-core/. In consequence a did validation with https://didlint.ownyourdata.eu/validate also fails with "invalid 'verificationMethod'". In addition if this issue is fixed the validator states that an additional context is missing: "@context: add 'https://w3id.org/security/suites/jws-2020/v1'"
Root cause: The three elements are missing in the did definition done on code level.
Solution: This MR addresses the three missing items by adding them on the code level. If this change is suitable for you, please consider fixing this also in the notary-registrationnumber component.
Fixes: no Jira issue

Signed-off-by: schuberr <robert.schubert@msg.group>