fix: get credential atomic types
Problem: types in json-ld can be set with various method
They can be put in the VerifiableCredential type array.
They can also be put in each credentialSubject type.
Types can be an array or a single string.
And finally, types can be in type or @type field

Cause: VerifiableCredentials spec is not enough constraint

Solution: Handle all cases, type & @type, array or not array

Fixes: #54

Signed-off-by: Ewann Gavard <ewann.gavard@gaia-x.eu>